+++ apis/python/src/tiledbsoma/_geometry_dataframe.py
@@ -298,6 +298,26 @@ class GeometryDataFrame(SpatialDataFrame, somacore.GeometryDataFrame):
         # if is it in read open mode, then it is a GeometryDataFrameWrapper
         return cast("GeometryDataFrameWrapper", self._handle).count
 
+    def delete_cells(
+        self,
+        coords: options.SparseDFCoords,
+        *,
+        value_filter: str | None = None,
+        platform_config: options.PlatformConfig | None = None,
+    ) -> None:
+        """Deletes cells at the specified coordinates.
+
+        Args:
+            coords:
+                A per-dimension ``Sequence`` of scalar, slice, sequence of scalar or
+                `Arrow IntegerArray <https://arrow.apache.org/docs/python/generated/pyarrow.IntegerArray.html>` values
+                defining the region to read.
+            value_filter:
+                An optional [value filter] to apply to the results.
+                Defaults to no filter.
+        """
+        raise NotImplementedError("Support for deleting cells from a geometry dataframe is not yet implemented.")
+
     def read(
         self,
         coords: options.SparseDFCoords = (),